session passthrough for the wasp-launcher
This is required to give wasp-launcher the ability to do sign-ins. Until now it only supported logging in, sign in was not possible because it requires admin database access which the wasp-launcher cannot have due to security reasons, so from now on the website will handle sign ins and log ins for the launcher and pass the session to it once it's done.
In a way similar to what the original wasp-launcher did before it has it's own embeded server.